refactor(worker): extract _toolset_ids_from_scoped helper (A4)
Move _toolset_ids_from_scoped to primer/worker/_toolset_ids.py and
re-export from primer.worker.pool (top-level, before the bottom
executor_builders import) so all importers keep resolving the old path.
Pure code motion, no behaviour change.
